move the sort function to a different file
